requirements.txt
aiohttp == 3.8.1
python-telegram-bot == 13.13
bech32 == 1.2.0
ecdsa == 0.18.0
hdwallets == 0.1.2
mnemonic == 0.20
python-dateutil == 2.8.2